// src/contexts/theme/types.ts
export interface IColors {
  0: React.CSSProperties["color"];
  5: React.CSSProperties["color"];
  10: React.CSSProperties["color"];
  20: React.CSSProperties["color"];
  30: React.CSSProperties["color"];
  40: React.CSSProperties["color"];
  50: React.CSSProperties["color"];
  60: React.CSSProperties["color"];
  70: React.CSSProperties["color"];
  80: React.CSSProperties["color"];
  90: React.CSSProperties["color"];
  100: React.CSSProperties["color"];
}

export interface IconColors {
  main: React.CSSProperties["color"];
  primary: React.CSSProperties["color"];
  secondary: React.CSSProperties["color"];
  positive: React.CSSProperties["color"];
  negative: React.CSSProperties["color"];
  warning: React.CSSProperties["color"];
  interactive: React.CSSProperties["color"];
  white: React.CSSProperties["color"];
}

export interface TertiaryColors {
  main: React.CSSProperties["color"];
  sand: React.CSSProperties["color"];
  dark: React.CSSProperties["color"];
  seal: React.CSSProperties["color"];
  darkBlue: React.CSSProperties["color"];
  lilacVoilet: React.CSSProperties["color"];
  darkOrange: React.CSSProperties["color"];
  chicLime: React.CSSProperties["color"];
  mediumSeaGreen: React.CSSProperties["color"];
  carolinaBlue: React.CSSProperties["color"];
  lighAquaGreen: React.CSSProperties["color"];
  mediumAquaGreen: React.CSSProperties["color"];
  indianRed: React.CSSProperties["color"];
  warmGray: React.CSSProperties["color"];
  darkPurple: React.CSSProperties["color"];
  glacierBlue: React.CSSProperties["color"];
  lightGray: React.CSSProperties["color"];
  mydocsGray: React.CSSProperties["color"];
  dimGray: React.CSSProperties["color"];
  interactiveGray: React.CSSProperties["color"];
}

export interface OtherColors {
  main: React.CSSProperties["color"];
  primaryBackground: React.CSSProperties["color"];
  secondaryBackground: React.CSSProperties["color"];
  darkBackground: React.CSSProperties["color"];
  divider: React.CSSProperties["color"];
  stroke: React.CSSProperties["color"];
  buttonRed: React.CSSProperties["color"];
}

export interface ComponentColors {
  surface: {
    dark: React.CSSProperties["color"];
    medium: React.CSSProperties["color"];
    light: React.CSSProperties["color"];
    white: React.CSSProperties["color"];
    scrim: React.CSSProperties["color"];
  };
  textIconOnLight: {
    primary: React.CSSProperties["color"];
    secondary: React.CSSProperties["color"];
    disabled: React.CSSProperties["color"];
    interactive: React.CSSProperties["color"];
  };
  textIconOnDark: {
    primary: React.CSSProperties["color"];
    secondary: React.CSSProperties["color"];
    disabled: React.CSSProperties["color"];
    interactive: React.CSSProperties["color"];
  };
  textStatusOnLight: {
    positive: React.CSSProperties["color"];
    poisitive2: React.CSSProperties["color"];
    information: React.CSSProperties["color"];
    success: React.CSSProperties["color"];
    warning: React.CSSProperties["color"];
    negative: React.CSSProperties["color"];
  };
  textStatusOnDark: {
    positive: React.CSSProperties["color"];
    poisitive2: React.CSSProperties["color"];
    information: React.CSSProperties["color"];
    success: React.CSSProperties["color"];
    warning: React.CSSProperties["color"];
    negative: React.CSSProperties["color"];
  };
  component: {
    primaryCTAEnabled: React.CSSProperties["color"];
    primaryCTADisabled: React.CSSProperties["color"];
    outlineEnabled: React.CSSProperties["color"];
    outlineDisabled: React.CSSProperties["color"];
    fillWhite: React.CSSProperties["color"];
    fillLight: React.CSSProperties["color"];
    fillPrefilled: React.CSSProperties["color"];
    dividerLight: React.CSSProperties["color"];
    dividerDark: React.CSSProperties["color"];
  };
  statusOnLight: {
    positive: React.CSSProperties["color"];
    positive2: React.CSSProperties["color"];
    information: React.CSSProperties["color"];
    success: React.CSSProperties["color"];
    warning: React.CSSProperties["color"];
    negative: React.CSSProperties["color"];
  };
  statusOnDark: {
    positive: React.CSSProperties["color"];
    positive2: React.CSSProperties["color"];
    information: React.CSSProperties["color"];
    success: React.CSSProperties["color"];
    warning: React.CSSProperties["color"];
    negative: React.CSSProperties["color"];
  };
  surfaceStatusLight: {
    positive: React.CSSProperties["color"];
    positive2: React.CSSProperties["color"];
    information: React.CSSProperties["color"];
    success: React.CSSProperties["color"];
    warning: React.CSSProperties["color"];
    negative: React.CSSProperties["color"];
  };
  surfaceStatusDark: {
    positive: React.CSSProperties["color"];
    positive2: React.CSSProperties["color"];
    information: React.CSSProperties["color"];
    success: React.CSSProperties["color"];
    warning: React.CSSProperties["color"];
    negative: React.CSSProperties["color"];
  };
  stateOnLight: {
    hover: React.CSSProperties["color"];
    focusPressed: React.CSSProperties["color"];
    selected: React.CSSProperties["color"];
    active: React.CSSProperties["color"];
    new: React.CSSProperties["color"];
    dragged: React.CSSProperties["color"];
  };
  stateOnDark: {
    hover: React.CSSProperties["color"];
    focusPressed: React.CSSProperties["color"];
    selected: React.CSSProperties["color"];
    active: React.CSSProperties["color"];
    new: React.CSSProperties["color"];
    dragged: React.CSSProperties["color"];
  };
  band: {
    eggPlant: React.CSSProperties["color"];
    warmGray: React.CSSProperties["color"];
    glacierBlue: React.CSSProperties["color"];
  };
}